+++ pages/notice.tsx
@@ -16,7 +16,7 @@ const Page: NextPageWithLayout = () => {
 
   return (
     <>
-      <Title backRoute="/home" title="공지사항 📣" subTitle="서비스 이용에 필요한 내용을 확인하세요" />
+      <Title backRoute="/" title="공지사항 📣" subTitle="서비스 이용에 필요한 내용을 확인하세요" />
       <Wrapper>
         <NoticeList>
           {noticeData.map((item: NoticeTypes, index: number) => {
