+++ store/user.ts
@@ -37,7 +37,7 @@ const userSlice = createSlice({
       state.kakaoUid = kakaoUid;
     },
     logoutAction: (state: UserTypes) => {
-      Router.push('/home'); // 메인화면으로 이동
+      Router.push('/'); // 메인화면으로 이동
       deleteToken(); // 헤더와 쿠키에서 토큰 제거
       state.isLoggedin = false;
       state.id = '';
